+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
 ----------
 
 - add migration snapshot script and release workflow
+- expand security model to include Public permissions and add permission matrix field
 
 0.1.7 [revision 4dcd62ed8222e102edbc88c81300a774e39f095c]
 ---------------------------------------------------------